update paper and waterfall
